(pthread_condattr_t): Size is only an int. (pthread_mutexattr_t): Likewise. (pthread_barrierattr_t): Likewise.

+++ nptl/sysdeps/unix/sysv/linux/powerpc/bits/pthreadtypes.h
@@ -78,7 +78,7 @@ typedef union
 typedef union
 {
   char __size[__SIZEOF_PTHREAD_MUTEXATTR_T];
-  long int __align;
+  int __align;
 } pthread_mutexattr_t;
 
 
@@ -101,7 +101,7 @@ typedef union
 typedef union
 {
   char __size[__SIZEOF_PTHREAD_CONDATTR_T];
-  long int __align;
+  int __align;
 } pthread_condattr_t;
 
 
@@ -171,7 +171,7 @@ typedef volatile int pthread_spinlock_t;
 typedef union
 {
   char __size[__SIZEOF_PTHREAD_BARRIER_T];
-  long int __align;
+  int __align;
 } pthread_barrier_t;
 
 typedef union
